Don't hardcode Int64 in tests

+++ test/ranges.jl
@@ -1300,7 +1300,7 @@ end
     @test convert(StepRange{Int128,Int128}, 0.:5) === Int128(0):Int128(1):Int128(5)
 
     @test StepRange(1:1:4) === 1:1:4
-    @test StepRange{Int32}(1:1:4) === StepRange{Int32,Int64}(1,1,4)
+    @test StepRange{Int32}(1:1:4) === StepRange{Int32,Int}(1,1,4)
 
     struct MyStepRange57718{T,S} <: OrdinalRange{T,S}
         r :: StepRange{T,S}
@@ -1311,7 +1311,7 @@ end
     Base.length(mr::MyStepRange57718) = length(mr.r)
 
     @test StepRange(MyStepRange57718(1:1:4)) === 1:1:4
-    @test StepRange{Int32}(MyStepRange57718(1:1:4)) === StepRange{Int32,Int64}(1,1,4)
+    @test StepRange{Int32}(MyStepRange57718(1:1:4)) === StepRange{Int32,Int}(1,1,4)
 
     @test_throws ArgumentError StepRange(1.1,1,5.1)
